Added App class
Added matchHandlers function to the class App
It provides flexibility to choose the handlers
according to the request

Added next concept to the class App
It perform the response task sequencially in predefined path

Commit Headline/ Feature

What changes you have made?
.....

Why this change is necessary?
....

What is the impact of the change?
.....